# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Microservice - api-core",
        "contact": {
            "email": "planeasy@marciodias.me"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://localhost:8000/",
            "description": "developer environment"
        }
    ],
    "paths": {
        "/domain/search": {
            "get": {
                "tags": [
                    "Domain"
                ],
                "summary": "Checks if the subdomain is available for registration",
                "description": "Check available subdomain to subdomain.planeasy.com.br",
                "operationId": "searchDomain",
                "parameters": [
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "Field Subdomain for planeasy.com.br",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get a list of available jobs to select in the register",
                "description": "List of jobs",
                "operationId": "getJobs",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get a list of available questions to select in the register",
                "description": "List of questions",
                "operationId": "getSignUpQuestions",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/signup/check-email": {
            "get": {
                "tags": [
                    "SignUp"
                ],
                "summary": "Check e-mail available to register",
                "description": "Verify e-mail available",
                "operationId": "checkEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Check e-mail available",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/signup/": {
            "post": {
                "tags": [
                    "SignUp"
                ],
                "summary": "Check e-mail available",
                "description": "Verify e-mail available",
                "operationId": "fe8366996a3cf4276e1710a669747c60",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "E-mail",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone mobile",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "job",
                        "in": "query",
                        "description": "Job UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_name",
                        "in": "query",
                        "description": "Name of company",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_phone_mobile",
                        "in": "query",
                        "description": "Phone of company",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_occupation_area",
                        "in": "query",
                        "description": "Company Occupation Area",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_employees_number",
                        "in": "query",
                        "description": "Company Employees",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "Company Employees",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "questions",
                        "in": "query",
                        "description": "Company Employees",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/workareas": {
            "get": {
                "tags": [
                    "WorkAreas"
                ],
                "summary": "Get a list of available work areas to select in the register",
                "description": "List of workareas",
                "operationId": "getSignUpWorkArea",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}